[PATCH v2] Cleanup: Use a negative number for uninitialized file descriptors
Do not use zero to denote an invalid file descriptor.

First, zero is a valid value, although quite unlikely to be used for
anything except standard input.

Second, open(2) returns a negative value on failure, so there are
already checks for a negative value in some places.

The purpose of this patch is not to allow running in an evil environment
(with closed stdin), but to aid in debugging by using a consistent value
for uninitialized file descriptors which is also regarded as invalid by
the kernel. For example, attempts to close a negative FD will fail
(unlike an attempt to close FD 0).

Changes from v1:
  - Cleanup file descriptor usage in dwarf_info.c and sadump_info.c

Signed-off-by: Petr Tesarik <ptesarik@suse.com>